+++ ddcommon/src/unix_utils/process.rs
@@ -6,8 +6,23 @@ use nix::sys::wait::{waitpid, WaitPidFlag, WaitStatus};
 use nix::unistd::Pid;
 use std::os::fd::RawFd;
 
-use super::errors::{PollError, ReapError};
 use super::timeout::TimeoutManager;
+
+#[derive(Debug, Eq, PartialEq, thiserror::Error)]
+pub enum ReapError {
+    #[error("Timeout waiting for child process to exit")]
+    Timeout,
+    #[error("Error waiting for child process to exit: {0}")]
+    WaitError(#[from] nix::Error),
+}
+
+#[derive(Debug, Eq, PartialEq, thiserror::Error)]
+pub enum PollError {
+    #[error("Poll failed with errno: {0}")]
+    PollError(i32),
+    #[error("Poll returned unexpected result: revents = {0}")]
+    UnexpectedResult(i16),
+}
 
 /// Non-blocking child reaper
 /// * If the child process has exited, return true
